+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.15.0 / 2026-07-30
+
+* [FEATURE] Add `-rewrite-host-header-to` flag to override the `Host` header in the requests forwarded to upstream (the proxy also adds the `X-Forwarded-Host` and `X-Forwarded-Proto` headers). #383
+* [ENHANCEMENT] Add `-log.format` and `-log.level` flags. #346
+
+## 0.14.0 / 2026-06-23
+
+* [FEATURE] Add the `-upstream-client-cert`, `-upstream-client-key` and `-upstream-server-name` flags to configure the TLS connection of the upstream server. #377
+
+## 0.13.0 / 2026-04-21
+
+* [FEATURE] Add the `-insecure-skip-verify` flag to bypass the TLS verification of the upstream server. #335
+* [FEATURE] Add the `-upstream-ca-cert` flag to provide the Certificate Authority's certificate of the upstream server. #340
+* [FEATURE] Add the `-enable-promql-extended-range-selectors` flag to support extended range selectors in PromQL expressions. #358
+* [FEATURE] Add the `-enable-promql-binop-fill-modifiers` flag to support binary operation fill modifiers in PromQL expressions. #358
+
+## 0.12.1 / 2025-09-11
+
+* [BUGFIX] Don't panic on `error-on-replace` with multiple values. #300
+
 ## 0.12.0 / 2025-08-06
 
 * [ENHANCEMENT] Add the `-enable-promql-duration-expression-parsing` flag to support arithmetic for durations in PromQL expressions. #297